consolidate code
